Fixed slow sync
