+++ opensrp-path/src/main/java/util/PathConstants.java
@@ -34,6 +34,9 @@ PathConstants extends AllConstants {
         public static final int WEIGHT_SYNC_PROCESSING = 6;
         public static final int RECURRING_SERVICES_SYNC_PROCESSING = 7;
         public static final int IMAGE_UPLOAD = 8;
+        public static final int STOCK_SYNC = 9;
+        public static final int ALERT_SYNC = 10;
+
     }
 
     public static final class EventType {
